third_party/blink/web_tests/external/wpt/webidl/ecmascript-binding/es-exceptions/DOMException-custom-bindings.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] If the implementation has a stack property on normal errors, it also does on DOMExceptions
  assert_equals: The typeof values must match expected "string" but got "undefined"
Harness: the test ran to completion.
